# internal/apiserver/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "IAM 系统 REST API 文档，包含认证(Authentication)、授权(Authorization)、身份管理(Identity)和身份提供商(IDP)模块",
        "title": "IAM API Documentation",
        "termsOfService": "https://iam.fangcunmount.cn/terms",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/FangcunMount/iam",
            "email": "support@fangcunmount.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "host": "iam.fangcunmount.cn",
    "basePath": "/api",
    "paths": {
...
        "/v2/authn/login-identities": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "列出当前用户登录身份",
                "responses": {
                    "200": {
                        "description": "当前用户已绑定的登录身份",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.LoginIdentityListResponse"
                        }
                    }
                }
            }
        },
        "/v2/authn/login-identities/phone": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "绑定手机号登录身份",
                "parameters": [
                    {
                        "description": "手机号与验证码",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkPhoneRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "绑定结果",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.LinkLoginIdentityResponse"
                        }
                    }
                }
            }
        },
        "/v2/authn/login-identities/phone/challenge": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "发送手机号绑定验证码",
                "parameters": [
                    {
                        "description": "手机号",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkPhoneChallengeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "已发送",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.MessageResponse"
                        }
                    }
                }
            }
        },
        "/v2/authn/login-identities/wechat-miniprogram": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "绑定微信小程序登录身份",
                "parameters": [
                    {
                        "description": "微信小程序 code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkWechatMiniProgramRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "绑定结果",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.LinkLoginIdentityResponse"
                        }
                    }
                }
            }
        },
        "/v2/authn/login-identities/wechat-open": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "完成微信开放平台扫码绑定",
                "parameters": [
                    {
                        "description": "微信回调 code 与 state",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkWechatOpenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "绑定结果",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.LinkLoginIdentityResponse"
                        }
                    }
                }
            }
        },
        "/v2/authn/login-identities/wechat-open/authorize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "发起微信开放平台扫码绑定授权",
                "parameters": [
                    {
                        "description": "可选 nonce",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkWechatOpenAuthorizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "授权地址与 state",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.WechatOpenAuthorizeResponse"
                        }
                    }
                }
            }
        },
        "/v2/authn/login-identities/wecom": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "绑定企业微信登录身份",
                "parameters": [
                    {
                        "description": "企业微信 code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkWecomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "绑定结果",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.LinkLoginIdentityResponse"
                        }
                    }
                }
            }
        },
        "/v2/authn/login-identities/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "解绑登录身份",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录身份 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "已解绑",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.MessageResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkPhoneChallengeRequest": {
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "phone": {
                    "type": "string"
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkPhoneRequest": {
            "type": "object",
            "required": [
                "otp_code",
                "phone"
            ],
            "properties": {
                "otp_code": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkWechatMiniProgramRequest": {
            "type": "object",
            "required": [
                "app_id",
                "code"
            ],
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkWechatOpenAuthorizeRequest": {
            "type": "object",
            "properties": {
                "nonce": {
                    "type": "string"
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkWechatOpenRequest": {
            "type": "object",
            "required": [
                "code",
                "state"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LinkWecomRequest": {
            "type": "object",
            "required": [
                "code",
                "corp_id"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "corp_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.LinkLoginIdentityResponse": {
            "type": "object",
            "properties": {
                "login_identity": {
                    "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.LoginIdentityResponse"
                },
                "reused": {
                    "type": "boolean"
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.LoginIdentityListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.LoginIdentityResponse"
                    }
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.LoginIdentityResponse": {
            "type": "object",
            "properties": {
                "global_identifier": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identifier": {
                    "type": "string"
                },
                "linked_at": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "realm": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "verified_at": {
                    "type": "string"
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.WechatOpenAuthorizeResponse": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string"
                },
                "authorize_url": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT 认证令牌，格式: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}